Flyttet Guesser ut av HomeController
